% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_watershed_cap.R
\name{calc_watershed_cap}
\alias{calc_watershed_cap}
\title{Watershed Capacity}
\usage{
calc_watershed_cap(
  wtsd_polygon,
  spp_range,
  capacity_pts,
  capacity_name = "chnk_per_m",
  capacity_se_name = "chnk_per_m_se",
  snap_dist = 500,
  by_stream = F
)
}
\arguments{
\item{wtsd_polygon}{sf polygon defining the extent you wish to estimate capacity over}

\item{spp_range}{sf line defining the species range extent}

\item{capacity_pts}{sf points with estimates of capacity}

\item{capacity_name}{character vector of the name of the capacity column in \code{capacity_pts}
#' @param capacity_name character vector of the name of the column in \code{capacity_pts} which characterizes the standard error of the capacity prediction}

\item{snap_dist}{numeric value describing maximum distance (in meters) to snap points to the species range extent. Default value is \code{500}.}

\item{by_stream}{Should capacities be returned by stream, \code{TRUE}, or just the total across the entire watershed, \code{FALSE}. Default is \code{FALSE}}
}
\description{
Estimate capacity within a polygon, based on line file of species domain and point estimates within the polygon
}
\author{
Kevin See
}
