% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowlines.R
\name{get_flowlines}
\alias{get_flowlines}
\title{Get Flowlines}
\usage{
get_flowlines(stream_order = 3, wtsd_polygon = NULL)
}
\arguments{
\item{stream_order}{maximum Strahler order of streams to return}

\item{wtsd_polygon}{sf polygon describing the boundry of where stream lines should be found}
}
\description{
Download flowlines with NHD flowline data for mapping purposes. I modified this from here: https://ryanpeek.github.io/2017-11-05-mapping-with-sf-Part-2/.
}
\author{
Kevin See
}
