% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{Impute Missing Data}
\usage{
impute_missing_data(
  data = NULL,
  covars = NULL,
  max_missing = 3,
  impute_vars = NULL,
  method = c("Hmisc", "missForest", "randomForestSRC"),
  ntree = 100,
  nk = 4,
  ...
)
}
\arguments{
\item{data}{data.frame containing \code{covars} and any \code{impute_vars} as columns}

\item{covars}{vector of covariate names to be imputed, which match column names in \code{data}}

\item{max_missing}{integer value of the maximum number of columns per row that may have missing values. Rows with more than this number of missing covariates will be deleted.}

\item{impute_vars}{additional covariates included in \code{data} that may be used for imputation}

\item{method}{which method to use, either by creating random forests using the \code{missForest} package, or by predictive mean matching using the \code{aregImpute} function in the \code{Hmisc} package}

\item{ntree}{how many trees to build if \code{method = 'missForest'} or \code{method == 'randomForestSRC'}}

\item{nk}{how many knots to use in smoothing splines if using \code{method = 'Hmisc'}}

\item{...}{other arguments to be passed to either \code{missForest}, \code{aregImpute} or \code{impute} functions}
}
\value{
data.frame containing non-missing and imputed data
}
\description{
Filter out rows with too much missing data, impute the remaining missing values
}
\author{
Kevin See
}
