% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapPointsToLines_v2.R
\name{snapPointsToLines_v2}
\alias{snapPointsToLines_v2}
\title{Snap Points to Lines}
\usage{
snapPointsToLines_v2(
  points,
  lines,
  maxDist = 0.1,
  withAttrs = TRUE,
  idField = NA
)
}
\arguments{
\item{points}{An object of the class SpatialPoints or SpatialPointsDataFrame.}

\item{lines}{An object of the class SpatialLines or SpatialLinesDataFrame.}

\item{maxDist}{Numeric value for establishing a maximum distance to avoid snapping points that are farther apart; its default value is NA.}

\item{withAttrs}{Boolean value for preserving (TRUE) or getting rid (FALSE) of the original point attributes. Default: TRUE. This parameter is optional.}

\item{idField}{A string specifying the field which contains each line's id. This id will be transferred to the snapped points data set to distinguish the line which each point was snapped to.}
}
\value{
SpatialPointsDataFrame object as defined by the R package 'sp'. This object contains the snapped points, therefore all of them lie on the lines.
}
\description{
This function snaps a set of points to a set of lines based on the minimum distance of each point to any of the lines. This function does not work with geographic coordinates. This function replaces the maptools::snapPointsToLines because it stopped working correctly
}
\author{
Kevin See
}
